test: 읽기 전용 API 분리 및 문서화 적용
